Rename files from notFoundEndpoint to notFounderror
