+++ src/server/error/notFoundError/__tests__/notFoundEndpoint.test.ts
